Rename `MultiTokenValueStat` -> `MultiValueStat`
