+++ packages/ui/src/common/components/statistics/index.tsx
@@ -4,6 +4,6 @@ export * from './StatisticHeader'
 export * from './StatisticItem'
 export * from './FractionValue'
 export * from './TokenValueStat'
-export * from './MultiTokenValueStat'
+export * from './MultiValueStat'
 export * from './DurationStatistics'
 export * from './StakeStat'
