+++ packages/ui/src/working-groups/components/MyEarningsStat.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 
-import { MultiTokenValueStat } from '@/common/components/statistics'
+import { MultiValueStat } from '@/common/components/statistics'
 import { useMyEarnings } from '@/working-groups/hooks/useMyEarnings'
 
 export const MyEarningsStat = () => {
   const earnings = useMyEarnings()
 
   return (
-    <MultiTokenValueStat
+    <MultiValueStat
       title="Earned in past"
       values={[
         { label: '24 hours', value: earnings?.last24hours },
